加上获取对象方法
